Mistakes in functions.

+++ preprocessing.py
@@ -24,7 +24,8 @@ def preprocessor():
     logging.info(f'Test-train split complete.')
 
     cat, con = cat_con_cols(X_train)  # Get the column names of the continuous and nominal data
-    X_train[cat] = X_train[cat].astype('Int64')  # Convert categorical cols values from floats to integers
+    X_train[cat] = X_train[cat].astype('Int64')  # Convert categorical cols values from floats to integers - train
+    X_test[cat] = X_test[cat].astype('Int64')  # Convert categorical cols values from floats to integers - test
 
     logging.info(f'X_train imputation started.')
     X_train_i4 = extra_trees_imputer(X_train[0:4000], cat, con)
